+++ src/ts/main.ts
@@ -106,6 +106,7 @@ export { CastUIContainer } from './components/CastUIContainer';
 export { PlaybackToggleOverlay, PlaybackToggleOverlayConfig } from './components/overlays/PlaybackToggleOverlay';
 export { CloseButton, CloseButtonConfig } from './components/buttons/CloseButton';
 export { BackButton } from './components/buttons/BackButton';
+export { RadioModeButton } from './components/buttons/RadioModeButton';
 export { MetadataLabel, MetadataLabelContent, MetadataLabelConfig } from './components/labels/MetadataLabel';
 export { AirPlayToggleButton } from './components/buttons/AirPlayToggleButton';
 export { VolumeSlider, VolumeSliderConfig } from './components/seekbar/VolumeSlider';